WIP: EdgeCluster class
this is currently an unfinished implementation that doesn't work
the edge clustering behaviour is exposed in python through the
Part.getSortedClusters function, so we'll use that for now. Eventually,
it would be nice to use our own code for this, so that the original impl
can be removed fro mteh main freecad codebase.